move purchase query to award files
