+++ Server/src/features/award/award.router.ts
@@ -1,6 +1,6 @@
 import { z } from 'zod';
 import { protectedProcedure, publicProcedure } from '../../core/trpc/trpc';
-import { awardService } from './award.service';
+import { awardService, purchase } from './award.service';
 import { router } from '../../core/trpc/trpc';
 
 export const awardRouter = router({
@@ -11,5 +11,10 @@ export const awardRouter = router({
     .input(z.object({ awardId: z.number() }))
     .mutation(async ({ input, ctx }) => {
       return await awardService.useAward(ctx.prisma ,ctx.userId, input.awardId);
+    }),
+  purchase: protectedProcedure
+    .input(z.object({ awardId: z.number() }))
+    .mutation(async ({ ctx, input }) => {
+      return await purchase(ctx.userId, input.awardId, ctx.prisma);
     }),
 });
