+++ Server/src/features/award/award.service.ts
@@ -37,3 +37,40 @@ export const awardService = {
     return { success: true };
   },
 };
+export const purchase = async (
+  userId: number,
+  awardId: number,
+  prisma: PrismaClient,
+): Promise<Purchase> => {
+  const user = await prisma.user.findUnique({ where: { id: userId } });
+  let award = await prisma.award.findUnique({ where: { id: awardId } }) as Award;
+
+  if (!user) throw new Error('User not found');
+  if (!award) throw new Error('Award not found');
+  if (user.coinBalance < award.price) throw new Error('Not enough coins');
+
+  const newPurchase = await prisma.purchase.create({
+    data: {
+      userId: user.id,
+      awardId: award.id,
+    },
+  });
+
+  await updateUserInfoAfterPurchase(award, user, prisma);
+
+  return newPurchase;
+};
+
+async function updateUserInfoAfterPurchase(
+  award: Award, user: User,
+  prisma: PrismaClient,
+) {
+
+  const data: any = { coinBalance: user.coinBalance - award.price };
+  if (award.category === AwardType.PROFILE_ICON) data['animalId'] = award.id;
+
+  await prisma.user.update({
+    where: { id: user.id },
+    data,
+  });
+}
